Move wait test to pilot/istioctl_test
